renamed "learn terstimonies" to "wrting effective testimonies"
